refactor: release note list 분리
