+++ src/containers/release-note/release-note-list/release-note-list.utils.ts
